really finish sign in screen
